Move TextInput stuff to uicommon package.
